Rely on collection testing module for common functions
